1. General refactoring 2. Remove HTML parse classes. 3. Add more options in settings. 4. Move to Android 2.3.3

+++ src/com/vmladenov/cook/core/converters/SimpleViewConverter.java
@@ -1,15 +1,16 @@
 package com.vmladenov.cook.core.converters;
 
 import android.database.Cursor;
+
 import com.vmladenov.cook.core.objects.SimpleView;
 
 public class SimpleViewConverter implements ICursorConverter<SimpleView> {
-    @Override
-    public SimpleView Convert(Cursor cursor) {
-        SimpleView simpleView = new SimpleView();
-        simpleView.id = cursor.getLong(0);
-        simpleView.title = cursor.getString(1);
-        return simpleView;
-    }
+	@Override
+	public SimpleView Convert(Cursor cursor) {
+		SimpleView simpleView = new SimpleView();
+		simpleView.id = cursor.getInt(0);
+		simpleView.title = cursor.getString(1);
+		return simpleView;
+	}
 
 }
